chore(index.html): add getUserList function to retrieve user list from local storage
The getUserList function is added to the index.html file. This function checks if the browser supports local storage and retrieves the user list from the local storage. If local storage is not supported, an empty array is returned. This function is useful for retrieving the user list for further processing or display.