feat(index.html): add form submission functionality to update user list with new user data
The changes in this commit add an event listener to the submit action of a form in the index.html file. When the form is submitted, the values of the input fields for name, age, and location are retrieved. These values are then used to create a new user object. The new user object is then passed to the putUserList function to update the user list. Finally, the renderUserList function is called to refresh the displayed user list. This functionality allows users to add new users to the list by submitting the form.